+++ example/index.vue
@@ -1,7 +1,10 @@
-<template>
-  <div class="container" style="color: var(--vue-color); background: var(--vue-bg);">
-    Vue example
-  </div>
+<template lang="ts">
+<div
+  class="container"
+  style="color: var(--vue-color); background: var(--vue-bg)"
+>
+  Vue example
+</div>;
 </template>
 
 <style>
@@ -13,5 +16,6 @@
 .container {
   padding: 1rem;
   border: 1px solid var(--vue-color);
+  color: var(--vue-bg);
 }
 </style>